bug 1787200: remove tecken from new eliot repository
We're splitting the Eliot project out from the Tecken project. One step
for that is to split the eliot code from the tecken repository.

This commit starts with a copy of the tecken repository, removes all the
tecken-specific material, and updates scripts to repurpose them for the
new eliot repository.