Update markus to 5.0.0, rework registered metrics, add eliot.pageview count
This updates markus to 5.0.0 which gives us RegisteredMetricsFilter.
That lets us remove our Eliot-bespoke RegisteredMetricsFilter. This also
moves metrics definition to eliot_metrics.yaml in a structure that's
(very roughly) similar to how Glean defines metrics.

When doing that, I discovered the health resource metrics were emitted
with a different MetricsInterface. I fixed that.

When doing that, I discovered we don't have a pageview incr like we do
with Socorro, so I implemented that. It emits a "eliot.pageview" incr
with the path and request method as tags. Because of the way Falcon does
routing, it's a little clunky.

I also fixed setup to set_up where it's a verb.